refactor: split file into 3 different files
